scripts: Skip deleting cert-manager for Cluster API provisioned clusters

+++ scripts/clean-sc.sh
@@ -13,6 +13,13 @@ fi
 
 here="$(dirname "$(readlink -f "$0")")"
 
+# shellcheck source=bin/common.bash
+source "${here}/../bin/common.bash"
+
+config_load sc
+
+clusterAPIEnabled=$(yq4 '.clusterApi.enabled' "${config[config_file_sc]}")
+
 GATE_VALWEBHOOK=$(
     "${here}/.././bin/ck8s" ops \
         kubectl sc get \
@@ -49,8 +56,12 @@ if [ -n "$CHALLENGES" ]; then
   done
 fi
 
-# Destroy cert-manager helm release
-"${here}/.././bin/ck8s" ops helmfile sc -l app=cert-manager destroy
+if [ "${clusterAPIEnabled}" = "false" ]; then
+  # Destroy cert-manager helm release
+  "${here}/.././bin/ck8s" ops helmfile sc -l app=cert-manager destroy
+else
+  log_info "Cluster API provisioned cluster, skipping deletion of cert-manager Helm release"
+fi
 
 # Destroy local-cluster minio release, otherwise pvc cleanup will get stuck
 helmfile -e local_cluster -f "${here}/../helmfile.d" -l app=minio destroy
@@ -61,9 +72,13 @@ helmfile -e local_cluster -f "${here}/../helmfile.d" -l app=minio destroy
 # Velero-specific removal: https://velero.io/docs/v1.10/uninstalling/
 "${here}/.././bin/ck8s" ops kubectl sc delete crds -l component=velero
 
-# Cert-manager specific removal
-"${here}/.././bin/ck8s" ops kubectl sc delete namespace cert-manager
-"${here}/.././bin/ck8s" ops kubectl sc delete crds -l app.kubernetes.io/name=cert-manager
+if [ "${clusterAPIEnabled}" = "false" ]; then
+  # Cert-manager specific removal
+  "${here}/.././bin/ck8s" ops kubectl sc delete namespace cert-manager
+  "${here}/.././bin/ck8s" ops kubectl sc delete crds -l app.kubernetes.io/name=cert-manager
+else
+  log_info "Cluster API provisioned cluster, skipping deletion of cert-manager namespace and CRDs"
+fi
 
 # Dex specific removal
 # Keep for now, we won't use Dex CRDs in the future
